LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-name-defining-rules-expected.txt


PASS @keyframes is unaffected by @scope
PASS @keyframes is unaffected by non-matching @scope
PASS @property is unaffected by @scope
PASS @property is unaffected by non-matching @scope
